requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
dj-database-url>=1.0.0